+++ restapi/repository/usersrepository.py
@@ -1,21 +1,11 @@
 from typing import AsyncGenerator
 
 from fastapi import Depends
-from fastapi_users.db import SQLAlchemyBaseUserTableUUID, SQLAlchemyUserDatabase
+from fastapi_users.db import SQLAlchemyUserDatabase
 from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker, create_async_engine
-from sqlalchemy.orm import DeclarativeBase
-from sqlalchemy import Column, String
+from models.user import User, Base
 import os
 
-# Modelos --------------------------------------------------
-class Base(DeclarativeBase):
-    pass
-
-class User(SQLAlchemyBaseUserTableUUID, Base):
-    name = Column(String)
-    surname1 = Column(String)
-    surname2 = Column(String)
-# ----------------------------------------------------------
 
 DATABASE = "sqlite+aiosqlite:///test.db"
